correct treat if loadRemoteURL fails

+++ hsapi/dbaccess/db_files.php
@@ -441,25 +441,29 @@ function downloadViaProxy($filename, $mimeType, $url, $bypassProxy = true){
     //if(!file_exists($filename)){ // || filemtime($filename)<time()-(86400*30))
 
     $rawdata = loadRemoteURLContent($url, $bypassProxy);
+    
+    if($rawdata!==false){
 
-    fileSave($rawdata, $filename);
-        /*
-        if ($raw) {
+        fileSave($rawdata, $filename);
+            /*
+            if ($raw) {
+
+            if(file_exists($filename)){
+            unlink($filename);
+            }
+            $fp = fopen($filename, "w");
+            //$fp = fopen($filename, "x");
+            fwrite($fp, $raw);
+            //fflush($fp);    // need to insert this line for proper output when tile is first requestet
+            fclose($fp);
+            }
+            */
+        //}
 
         if(file_exists($filename)){
-        unlink($filename);
+            downloadFile($mimeType, $filename);
         }
-        $fp = fopen($filename, "w");
-        //$fp = fopen($filename, "x");
-        fwrite($fp, $raw);
-        //fflush($fp);    // need to insert this line for proper output when tile is first requestet
-        fclose($fp);
-        }
-        */
-    //}
-
-    if(file_exists($filename)){
-        downloadFile($mimeType, $filename);
+    
     }
 }
